test : [백] 행사 검색 관련 Test 추가 및 기타 Test 수정
