+++ src/businessservice/src/test/java/com/eventty/businessservice/application/EventServiceTest.java
@@ -3,7 +3,7 @@ package com.eventty.businessservice.application;
 import com.eventty.businessservice.domains.event.application.dto.request.EventUpdateRequestDTO;
 import com.eventty.businessservice.domains.event.application.dto.response.EventWithTicketsFindByIdResponseDTO;
 import com.eventty.businessservice.domains.event.application.dto.response.EventBasicFindAllResponseDTO;
-import com.eventty.businessservice.domains.event.application.Facade.EventService;
+import com.eventty.businessservice.domains.event.application.service.EventService;
 import com.eventty.businessservice.domains.event.domain.entity.EventDetailEntity;
 import com.eventty.businessservice.domains.event.domain.entity.EventBasicEntity;
 import com.eventty.businessservice.domains.event.domain.entity.TicketEntity;
@@ -32,48 +32,9 @@ public class EventServiceTest {
 
     @Mock
     private EventBasicRepository eventBasicRepository;
-    @Mock
-    private EventDetailRepository eventDetailRepository;
-
-    @Mock
-    private TicketRepository ticketRepository;
 
     @InjectMocks
     private EventService eventService;
-
-    @Test
-    @DisplayName("존재하는 이벤트 조회 테스트")
-    public void findEventByIdTest_ExistingEvent() {
-        // Given
-        Long eventId = 1L;
-        EventFullFindByIdResponseDTO mockEvent = createEventWithDetailDAO(eventId);
-        List<TicketEntity> mockTickets = createTickets(); // 필요에 따라 만들어야 합니다.
-
-        when(eventBasicRepository.selectEventWithDetailById(eventId)).thenReturn(mockEvent);
-        when(ticketRepository.selectTicketByEventId(eventId)).thenReturn(mockTickets);
-
-        // When
-        EventWithTicketsFindByIdResponseDTO responseDTO = eventService.findEventById(eventId);
-
-        // Then
-        assertEquals(mockEvent.getId(), responseDTO.getId());
-        assertEquals(mockEvent.getTitle(), responseDTO.getTitle());
-        assertEquals(mockEvent.getContent(), responseDTO.getContent());
-
-        verify(eventBasicRepository, times(1)).selectEventWithDetailById(eventId);
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 이벤트 조회 시 예외 테스트")
-    public void findEventByIdTest_NonExistingEvent() {
-        // Given
-        Long eventId = 1L;
-        when(eventBasicRepository.selectEventWithDetailById(eventId)).thenReturn(null);
-
-        // When & Then
-        assertThrows(EventNotFoundException.class, () -> eventService.findEventById(eventId));
-        verify(eventBasicRepository, times(1)).selectEventWithDetailById(eventId);
-    }
 
     @Test
     @DisplayName("전체 이벤트 조회 테스트")
@@ -91,54 +52,11 @@ public class EventServiceTest {
     }
 
     @Test
-    @DisplayName("이벤트 삭제 테스트")
-    public void deleteEventTest(){
-        // Given
-        Long eventId = 1L;
-        when(eventBasicRepository.deleteEvent(eventId)).thenReturn(1L);
-        when(eventDetailRepository.deleteEventDetail(eventId)).thenReturn(1L);
-        when(ticketRepository.deleteTicket(eventId)).thenReturn(1L);
-
-        // When
-        eventService.deleteEvent(eventId);
-
-        // Then
-        assertNull(eventBasicRepository.selectEventById(eventId));
-        verify(eventBasicRepository, times(1)).deleteEvent(eventId);
-        verify(eventDetailRepository, times(1)).deleteEventDetail(eventId);
-    }
-
-    @Test
-    @DisplayName("이벤트 수정 테스트")
-    public void updateEventTest() {
-        Long eventId = 1L;
-        EventBasicEntity existingEvent = createEventEntity(eventId);
-
-        EventUpdateRequestDTO updateRequestDTO = EventUpdateRequestDTO.builder()
-                .title("Updated Title")
-                .image("updated_image.jpg")
-                .content("Updated Content")
-                .build();
-
-        EventDetailEntity existingEventDetail = EventDetailEntity.builder()
-                .id(eventId)
-                .content("Old Content")
-                .build();
-
-        when(eventBasicRepository.selectEventById(eventId)).thenReturn(existingEvent);
-        when(eventDetailRepository.selectEventDetailById(eventId)).thenReturn(existingEventDetail);
-
-        eventService.updateEvent(eventId, updateRequestDTO);
-
-        verify(eventBasicRepository, times(1)).updateEvent(any(EventBasicEntity.class));
-        verify(eventDetailRepository, times(1)).updateEventDetail(any(EventDetailEntity.class));
-    }
-
-    @Test
+    @DisplayName("이벤트 카테고리별 조회 테스트")
     public void findEventsByCategoryTest() {
         // given
-        Long categoryId = 5L;
-        List<EventBasicEntity> mockEvents = createEventEntityList(3L); // 필요에 따라 만들어야 합니다.
+        Long categoryId = 1L;
+        List<EventBasicEntity> mockEvents = createEventEntityList(3L);
         when(eventBasicRepository.selectEventsByCategory(categoryId)).thenReturn(mockEvents);
 
         // when
@@ -151,6 +69,7 @@ public class EventServiceTest {
     }
 
     @Test
+    @DisplayName("이벤트 카테고리별 조회 테스트 - 존재하지 않는 카테고리")
     public void findEventsByCategoryTest_InvalidCategoryId() {
         // given
         Long invalidCategoryId = 15L;
@@ -162,6 +81,7 @@ public class EventServiceTest {
     }
 
     @Test
+    @DisplayName("이벤트 카테고리별 조회 테스트 - 이벤트 없음")
     public void findEventsByCategoryTest_NoEventsFound() {
         // given
         Long categoryId = 7L;
@@ -171,6 +91,23 @@ public class EventServiceTest {
         assertThrows(EventNotFoundException.class, () -> {
             eventService.findEventsByCategory(categoryId);
         });
+    }
+
+    @Test
+    @DisplayName("이벤트 검색 테스트")
+    public void findEventsBySearchTest() {
+        // given
+        String keyword = "Sample";
+        List<EventBasicEntity> mockEvents = createEventEntityList(3L);
+        when(eventBasicRepository.selectEventsBySearch(keyword)).thenReturn(mockEvents);
+
+        // when
+        List<EventBasicFindAllResponseDTO> result = eventService.findEventsBySearch(keyword);
+
+        // then
+        assertNotNull(result);
+        assertFalse(result.isEmpty());
+        assertEquals(mockEvents.size(), result.size());
     }
 
     private static EventBasicEntity createEventEntity(Long i){
@@ -189,26 +126,6 @@ public class EventServiceTest {
             .build();
     }
 
-    private static EventFullFindByIdResponseDTO createEventWithDetailDAO(Long id){
-        return EventFullFindByIdResponseDTO.builder()
-                .id(id)
-                .userId(1L)
-                .title("Sample Event")
-                .image("sample.jpg")
-                .eventStartAt(Timestamp.valueOf("2023-08-21 10:00:00").toLocalDateTime())
-                .eventEndAt(Timestamp.valueOf("2023-08-21 15:00:00").toLocalDateTime())
-                .participateNum(100L)
-                .location("Sample Location")
-                .category(1L)
-                .isActive(true)
-                .isDeleted(false)
-                .content("Sample content")
-                .applyStartAt(Timestamp.valueOf("2023-08-21 10:00:00").toLocalDateTime())
-                .applyEndAt(Timestamp.valueOf("2023-08-21 15:00:00").toLocalDateTime())
-                .views(100L)
-                .build();
-    }
-
     private static List<EventBasicEntity> createEventEntityList(Long count) {
         List<EventBasicEntity> eventBasicEntityList = new ArrayList<>();
 
@@ -219,35 +136,5 @@ public class EventServiceTest {
 
         return eventBasicEntityList;
     }
-
-    private List<TicketEntity> createTickets() {
-        List<TicketEntity> tickets = new ArrayList<>();
-
-        // 티켓 1
-        TicketEntity ticket1 = TicketEntity.builder()
-                .id(1L)
-                .name("Ticket 1")
-                .price(2000L)
-                .quantity(50L)
-                .eventId(1L)
-                .is_deleted(false)
-                .build();
-
-        // 티켓 2
-        TicketEntity ticket2 = TicketEntity.builder()
-                .id(2L)
-                .name("Ticket 2")
-                .price(3000L)
-                .quantity(30L)
-                .eventId(1L)
-                .is_deleted(false)
-                .build();
-
-        tickets.add(ticket1);
-        tickets.add(ticket2);
-
-        return tickets;
-    }
-
 
 }
